Restructure OutputPreview
