+++ src/ui/ide/OutputPreview/FolderPanel/FolderItem.tsx
