+++ src/ui/ide/OutputPreview/PathBar/PathItem.tsx
